+++ activities/FractionBounce.activity/js/activity.js
@@ -31,6 +31,7 @@ let app = new Vue({
 		traction: 0.1,
 		paused: true,
 		onSlope: true,
+		SugarL10n: null,
 
 		frameInterval: 20,
 		launchDelay: 1000,
@@ -44,6 +45,8 @@ let app = new Vue({
 		userFractions: [],
 		successSound: null,
 		failSound: null,
+		score: 0,
+		count: 0,
 		l10n: {
 			stringFractionBounceActivity: '',
 			stringTemplate: '',
@@ -123,6 +126,8 @@ let app = new Vue({
 		if(this.$refs.SugarDevice.isMobile()) {
 			this.$refs.SugarDevice.watchAcceleration(2*this.frameInterval);
 		}
+
+		this.SugarL10n = this.$refs.SugarL10n;
 	},
 
 	methods: {
@@ -225,6 +230,11 @@ let app = new Vue({
 			this.paused = !this.paused;
 			if (!this.paused) {
 				this.launch();
+				if(this.count==30){
+					this.score = 0;
+					this.count = 0;
+			        Score.innerHTML = " ";	
+				}
 			}
 			document.getElementById('slopeCanvas').removeEventListener('click', this.startGame);
 		},
@@ -295,10 +305,15 @@ let app = new Vue({
 				slopeCanvas.removeEventListener("mousedown", this.onTouchStart);
 				this.onTouchEnd();
 				let result = this.$refs.slopecanvas.checkAnswer();
+				this.count++;
 				if (result == this.answer) {
 					this.successSound.play();
+					this.score++;
 				} else {
 					this.failSound.play();
+				}
+				if (this.count==30) {
+					Score.innerHTML = this.SugarL10n.get("Score")+" "+ Math.round(this.score / 3) + "/10";
 				}
 				this.onSlope = true;
 				if (this.vy < 0.5) {